SiemAgent/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# src\util\config_loader.py: 1
PyYAML == 3.12

# src\util\utc_converter.py: 1,3
python_dateutil == 2.7.2

# src\util\utc_converter.py: 2
pytz == 2018.4

# src\reader\windows_log_reader.py: 1,2
pywin32 == 223

# src\service\siem_service.py: 4
requests_pkcs12 == 1.0.3

# src\reader\file_log_reader.py: 6,7,8
watchdog == 0.8.3
